+++ README.md
@@ -1,2 +1,7 @@
 # Personal Finance App
 A ReactJS/Python Flask Personal Finance App for HackThis2020
+
+# Api
+Start the API server by running:
+
+    python start_server.py
