+++ api/models/__init__.py
@@ -1,0 +1,5 @@
+from .db_models import db
+
+# Initialize sqlalchemy db backend
+def init_app(app):
+    db.init_app(app)
