+++ start_server.py
@@ -1,0 +1,4 @@
+import api
+
+app = api.create_app()
+app.run()
